fixed GearScore (formula)

+++ src/server/game/Entities/Player/Player.h
@@ -2216,7 +2216,7 @@ class TC_GAME_API Player : public Unit, public GridObject<Player>
         Spell* m_spellModTakingSpell;
 
         float GetAverageItemLevel() const;
-        float GetGearScore() const;
+        uint32 GetGearScore() const;
         bool isDebugAreaTriggers;
 
         void ClearWhisperWhiteList() { WhisperList.clear(); }
